Social Network
This is a small programe which open and read two files. One file is a
list with a couple of names and another one is a paragrph.  This
programe will search each names from the name list which cooccurence
within 5 lines one by one and write  all the cooccurence in a .csv file.
This programe used structure array implements linked list in C. Also
file stream is also used for file read and write.